Improve fling gestures on iOS (#2851)
Move `WebVelocityTracker1D` to the skiko source set and rename it to the
`PointerVelocityTracker1D`.
Add `preventReversedPointerMovements` option that prevents velocity
tracker from returning velocity with the opposite direction to the
general scroll direction.

Fixes
https://youtrack.jetbrains.com/issue/CMP-9297/Fling-gestures-not-working-correctly-in-LazyColumn-on-iOS.

## Release Notes
### Fixes - iOS
- Fix the scrolling inertia issue when performing short scroll gestures.
- Fix an issue where a fling may occur unexpectedly when lifting a
finger.

---------
